requirements.txt
asgiref==3.8.1
Django==5.2.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
